pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Agent-agent"
version = "0.1.0"
description = "Autonomous Systems â€” 3-Layer Agent Architecture"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }

dependencies = [
    # AI / LLM
    "openai>=1.30",
    # HTTP
    "httpx>=0.27",
    # Configuration & secrets
    "python-dotenv>=1.0",
    # Data validation & schemas
    "pydantic>=2.7",
    # Google APIs (Sheets, Drive)
    "google-auth>=2.29",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.130",

    # Bluesky / AT Protocol
    "atproto>=0.0.54",
    # Rich CLI output
    "rich>=13.7",
    "typer>=0.12",
    # Structured logging
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "black>=24.4",
    "mypy>=1.10",
]

[project.scripts]
Agent = "main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["orchestrator*", "executor*"]


[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=orchestrator --cov=executor --cov-report=term-missing"

